Move color to its own package
